+++ kernel/arch/dreamcast/hardware/pvr/pvr_scene.c
@@ -9,6 +9,7 @@
 #include <assert.h>
 #include <stdio.h>
 #include <string.h>
+#include <kos/genwait.h>
 #include <kos/string.h>
 #include <kos/thread.h>
 #include <dc/pvr.h>
@@ -86,10 +87,24 @@ void pvr_vertbuf_written(pvr_list_t list, uint32 amt) {
     pvr_state.dma_buffers[pvr_state.ram_target].ptr[list] = val;
 }
 
+static void pvr_start_ta_rendering(void) {
+    // Make sure to wait until the TA is ready to start rendering a new scene
+    if(!pvr_state.ta_ready) {
+        pvr_wait_ready();
+        pvr_state.ta_ready = 1;
+    }
+
+    // Starting from that point, we consider that the Tile Accelerator
+    // might be busy.
+    pvr_state.ta_busy = 1;
+}
+
 /* Begin collecting data for a frame of 3D output to the off-screen
    frame buffer */
 void pvr_scene_begin(void) {
     int i;
+
+    pvr_state.ta_ready = 0;
 
     // Get general stuff ready.
     pvr_state.list_reg_open = -1;
@@ -164,8 +179,10 @@ int pvr_list_begin(pvr_list_t list) {
 
     pvr_list_dma = pvr_list_uses_dma(list);
 
-    if(!pvr_list_dma)
+    if(!pvr_list_dma) {
+        pvr_start_ta_rendering();
         sq_lock((void *)PVR_TA_INPUT);
+    }
 
     /* Ok, set the flag */
     pvr_state.list_reg_open = list;
@@ -304,12 +321,21 @@ int pvr_scene_finish(void) {
         b = pvr_state.dma_buffers + pvr_state.ram_target;
 
         for(i = 0; i < PVR_OPB_COUNT; i++) {
-            /* Check whether the current list type should be skipped:
-               A. We never enabled the list globally with pvr_init().
-               B. We never associated an in-RAM DMA vertex buffer with
-                  the given list type, because we're using hybrid
-                  rendering and submitted that list type directly. */
-            if(!(pvr_state.lists_enabled & (1 << i)) || !b->base[i])
+            /* We never enabled the list globally with pvr_init() - skip it */
+            if(!(pvr_state.lists_enabled & (1 << i)))
+                continue;
+
+            /* If any lists weren't used in this scene, submit blank ones now */
+            if(!(pvr_state.lists_closed & (1 << i))) {
+                pvr_list_begin(i);
+                pvr_blank_polyhdr(i);
+                pvr_list_finish();
+            }
+
+            /* We never associated an in-RAM DMA vertex buffer with the given
+               list type, because we're using hybrid rendering and submitted
+               that list type directly - skip it */
+            if(!b->base[i])
                 continue;
 
             // Make sure there's at least one primitive in each.
@@ -326,6 +352,8 @@ int pvr_scene_finish(void) {
             assert(b->ptr[i] <= b->size[i]);
         }
 
+        pvr_start_ta_rendering();
+
         // Flip buffers and mark them complete.
         o = irq_disable();
         pvr_state.dma_buffers[pvr_state.ram_target].ready = 1;
@@ -333,6 +361,8 @@ int pvr_scene_finish(void) {
         irq_restore(o);
 
         pvr_sync_stats(PVR_SYNC_BUFDONE);
+
+        pvr_start_dma();
     }
     else {
         /* If a list was open, close it */
@@ -355,11 +385,16 @@ int pvr_scene_finish(void) {
 }
 
 int pvr_wait_ready(void) {
-    int t;
+    int flags, t = 0;
 
     assert(pvr_state.valid);
 
-    t = sem_wait_timed((semaphore_t *)&pvr_state.ready_sem, 100);
+    flags = irq_disable();
+
+    if(pvr_state.ta_busy)
+        t = genwait_wait((void *)&pvr_state.ta_busy, "PVR wait ready", 100, NULL);
+
+    irq_restore(flags);
 
     if(t < 0) {
 #if 0
@@ -383,7 +418,7 @@ int pvr_wait_ready(void) {
 int pvr_check_ready(void) {
     assert(pvr_state.valid);
 
-    if(sem_count((semaphore_t *)&pvr_state.ready_sem) > 0)
+    if(!pvr_state.ta_busy)
         return 0;
     else
         return -1;